backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
pydantic==2.10.0
httpx==0.27.2
scikit-learn==1.5.2
google-generativeai==0.8.3